# backend/openapi.yml
openapi: 3.0.0
info:
  title: Multi-Tenant SaaS Platform API
  version: 1.0.0
  description: >-
    API documentation for the Multi-Tenant SaaS Platform with Project & Task
    Management. Fully complies with the 19 API endpoints requirement and the
    {success, message, data} uniform response format.
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Unauthorized (Token missing or invalid)
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Unauthorized access
    Forbidden:
      description: Forbidden (Insufficient permissions)
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: 'Forbidden: You do not have permission'
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Resource not found
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      summary: 2. Login to the system
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: admin@demo.com
                password:
                  type: string
                  example: Demo@123
                tenantSubdomain:
                  type: string
                  example: demo
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                            example: bbbbbbbb-bbbb-bbbb
                          email:
                            type: string
                            example: admin@demo.com
                          fullName:
                            type: string
                            example: Demo Admin
                          role:
                            type: string
                            example: tenant_admin
                          tenantId:
                            type: string
                            format: uuid
                            example: 22222222-2222-2222
                      token:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      expiresIn:
                        type: string
                        example: 86400
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tenants:
    get:
      summary: 7. List all tenants (Super Admin Only)
      tags:
        - Tenants
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      tenants:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: '1111'
                            name:
                              type: string
                              example: Demo Company
                            subdomain:
                              type: string
                              example: demo
                            status:
                              type: string
                              example: active
                            subscriptionPlan:
                              type: string
                              example: professional
                            totalUsers:
                              type: number
                              example: 3
                            totalProjects:
                              type: number
                              example: 2
                            createdAt:
                              type: timestamp
                              example: '2024-12-26T10:00:00.000Z'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /projects:
    post:
      summary: 12. Create Project
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Swagger Test Project
                description:
                  type: string
                  example: This project was created via Swagger testing
                status:
                  type: enum
                  enum: ["active", "archieved", "completed"]
                  example: active
      responses:
        '201':
          description: Project created 
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Project created successfully
                  data:
                    type: object
                    properties:
                      id:
                        type: uuid
                        example: project-uuid
                      tenantId:
                        type: uuid
                        example: tenant-uuid
                      name:
                        type: string
                        example: Swagger Test Project
                      description:
                        type: string
                        example: This project was created via Swagger testing
                      status:
                        type: enum
                        enum: ["active", "archieved", "completed"]
                        example: active
                      createdBy:
                        type: uuid
                        example: creator-uuid
                      createdAt:
                        type: timestamp
                        example: 2024-12-27T08:00:00.000Z
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      summary: 13. List Projects
      tags:
        - Projects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      projects:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: uuid
                              example: project-uuid
                            name:
                              type: string
                              example: Swagger Test Project
                            description:
                              type: string
                              example: This project was created via Swagger testing
                            status:
                              type: string
                              example: active
                            createdBy:
                              type: object
                              properties:
                                id:
                                  type: uuid
                                  example: creator-uuid
                                fullName:
                                  type: string
                                  example: Demo Admin
                            taskCount:
                              type: integer
                              example: 5
                            completedTaskCount:
                              type: integer
                              example: 2
                            createdAt:
                              type: timestamp
                              example: 2024-12-27T08:00:00.000Z
                      total:
                        type: integer
                        example: 2
                      pagination:
                        type: object
                        properties:
                          currentPage:
                            type: integer
                            example: 1
                          totalPages:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 20
...